requirements.txt
pillow==7.2.0
numpy>=1.19.0
opencv-python==4.3.0.36
